Flesh out logic.
