--- public/source/app.css
@@ -1,45 +1,0 @@
-cf-app[app-id="a-better-browser"] {
-  background: #45484d;
-  position: absolute;
-  z-index: 100000;
-  width: 100%;
-  height: 26px;
-  top: 0;
-  left: 0;
-  overflow: hidden;
-  padding: 8px 0;
-  font: 18px/26px Arial,sans-serif;
-  text-align: center;
-  color: #FFF;
-  box-sizing: content-box
-}
-
-cf-app[app-id="a-better-browser"] a {
-  text-decoration: underline;
-  color: #EBEBF4;
-}
-
-cf-app[app-id="a-better-browser"] a:hover,
-cf-app[app-id="a-better-browser"] a:active {
-  text-decoration: none;
-  color: #DBDBEB;
-}
-
-cf-app[app-id="a-better-browser"] .close {
-  background: #393b40;
-  display: block;
-  width: 42px;
-  height: 42px;
-  position: absolute;
-  text-decoration: none !important;
-  cursor: pointer;
-  top: 0;
-  right: 0;
-  font-size: 30px;
-  line-height: 42px;
-}
-
-cf-app[app-id="a-better-browser"] .close:hover {
-  background: #E04343;
-  color: #FFF;
-}
